# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation for Dihauti Residence",
        "description": "description",
        "contact": {
            "email": "dev.arven@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.dihauti.ru/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/posts/{blogPost}/comments": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get all comment for post with pagination",
                "description": "Returns all comment for post with pagination",
                "operationId": "getCommentList",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comments for blog post retrieved successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Create yet another comment for post in database",
                "operationId": "commentCreate",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comment store successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
        "/posts/{blogPost}/comments/{id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get comment for post by id",
                "description": "Returns comment for post by id",
                "operationId": "getCommentById",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment for post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment for blog post retrieved successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Comment"
                ],
                "summary": "Update comment for post in database",
                "operationId": "commentUpdate",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment for post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Comment updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Delete comment for post in database",
                "operationId": "commentDelete",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Post slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment for post id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Comment deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Password Based": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "flows": {
                    "password": {
                        "authorizationUrl": "/login",
                        "refreshUrl": "/logout",
                        "scopes": []
                    }
                },
                "scheme": "https"
            }
        }
    }
}